update openAPI endpoints

+++ Travel Schedule/Services/NearestSettlementService.swift
@@ -12,7 +12,7 @@ import OpenAPIURLSession
 typealias Settlement = Components.Schemas.Settlement
 
 protocol NearestSettlementServiceProtocol {
-    func getNearestSettlement(lat: Double, lng: Double, distance: Int) async throws -> Settlement
+    func getNearestSettlement(lat: Double, lng: Double, distance: Double) async throws -> Settlement
 }
 
 final class NearestSettlementService: NearestSettlementServiceProtocol {
@@ -22,7 +22,7 @@ final class NearestSettlementService: NearestSettlementServiceProtocol {
         self.client = client
     }
 
-    func getNearestSettlement(lat: Double, lng: Double, distance: Int) async throws -> Settlement {
+    func getNearestSettlement(lat: Double, lng: Double, distance: Double) async throws -> Settlement {
         let response = try await client.getNearestSettlement(
             query: .init(
                 lat: lat,
